FILE: renames get-options-by-expansion files
