+++ libs/spirit-islander/shared/util/src/lib/functions/get-options.ts
