+++ libs/spirit-islander/shared/util/src/lib/functions/index.ts
@@ -1,7 +1,7 @@
 export * from './get-adversary-level-ids';
 export * from './get-difficulty';
 export * from './get-names';
-export * from './get-options-by-expansion';
+export * from './get-options';
 export * from './get-options-by-name';
 export * from './get-spirits-by-expansion';
 export * from './group-spirits';
